buddy list: Add links below sections to see full user list.
We don't show all users in the buddy list, and
specifically exclude some users that have been
inactive for a while. To see all subscribed
users, you can now click on a link below the
subscribed users section. The "other" section
shows a link to all users in the organization.

Related CZO conversation:
https://chat.zulip.org/#narrow/stream/101-design/topic/buddy.20list.20split.20users/near/1669158